pyproject.toml
[project]
name = "proctoring-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.2",
    "uvicorn>=0.30.6",
    "opencv-python>=4.6.0.66",
    "ultralytics>=8.3.251",
    "mediapipe>=0.10.21",
    "deepgram-sdk>=5.3.1",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.1",
    "numpy>=1.26.4",
    "pyannote.audio>=4.0.3",
    "sounddevice>=0.5.3",
]
